Created file 'Worlds'
This page has information and images on/of all fifteen worlds. As I could not add images to the Github editor, I leave the images in link form at the bottom of the page. Though I am aware that a 'Planets' file already exists, it only contains Arbusto, and not much information on it at that. This is a completed, up-to-date list which I collaborated on with ArjunCraft2001.